requirements.txt
# Web 框架
streamlit==1.31.0

# HTTP 客户端
httpx==0.27.0
tenacity==8.2.3

# 数据验证
pydantic==2.6.0
pydantic-settings==2.1.0

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1

# 模板引擎
jinja2==3.1.3

# 日志
loguru==0.7.2

# 图像处理
pillow==10.2.0

# 进度条
tqdm==4.66.1

# 测试
pytest==8.0.0
pytest-asyncio==0.23.4
